feat(internet-search): add tests for compact_search_result function to ensure relevant fields are retained
