+++ apps/desktop/src-tauri/src/functions/ai/copilot_sdk_tools.rs
@@ -2316,24 +2316,4 @@ mod tests {
             Some("internet_search")
         );
     }
-
-    #[test]
-    fn compact_search_result_keeps_only_model_relevant_fields() {
-        let result = compact_search_result(&json!({
-            "title": "Flow Like",
-            "url": "https://flow-like.com",
-            "content": "Workflow automation",
-            "publishedDate": "2026-06-04",
-            "engine": "test",
-            "category": "general",
-            "score": 1.25,
-            "huge": "drop me"
-        }));
-
-        assert_eq!(
-            result.get("title").and_then(Value::as_str),
-            Some("Flow Like")
-        );
-        assert!(result.get("huge").is_none());
-    }
 }
